PP-2117: Add new API to support streaming queries in Vizier.
Summary:
This is just the skeleton of the API to make sure we are all on board with the api before I implement the changes.
We want to use the same message to stream data from PEMs->Kelvins as from Kelvins->Query brokers. This message should include the ability to send execution stats for the query, since those will be streamed periodically for the streaming queries. This API will subsume the old API, the difference with batch queries will be that they terminate and send end of stream, where as streaming queries will go on until they are cancelled. Next up will be to refactor GRPCSink/Source and GRPC router to use this new API. After that point, a GRPC router will be implemented in go (highly based off of the c++ version we already have) in order to receive streamed query results from Kelvin nodes.

Test Plan: n/a

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5919

GitOrigin-RevId: 7b06ac56d5397bd56f7d20cacbd03435e0ae1e6c