Fix updater image path for public images
Summary:
we recently updated the build so that it pushes the images to a public repo for official releases...
unfortunately, there was a hard-coded path for the updater image hidden in our files.
this broke 0.4.0, but i put out a temporary fix by uploading the correct image to the pl-dev-infra registry for now.

this diff takes the image path from the cloudconnector image and uses that image path for the updater image.

Test Plan: ran skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5926

GitOrigin-RevId: 23b3641dec96ec69440f388c14729a9a314c0e6c