PP-2117: Replace TransferRowBatch API with TransferResultChunk
Summary:
TransferResultChunk is the new API for streaming row batches to their recipients. It supports sending query stats as well as sending result data to non-Carnot sources, such as the query broker. This change will allow GRPCSinks to send directly to the query broker, which will support streaming mode and also remove the need for Kelvins to buffer up the data before sending it to the query broker (even in batch mode, this will be good because it will allow results to show up as they are available).

Changed the field names on some protobufs so that the terminology and naming would be more consistent and less confusing going forward. This should be an ok set of changes because it is all internal to Vizier.

Test Plan: existing tests

Reviewers: zasgar, michelle, philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5927

GitOrigin-RevId: acf29b492312a0f9780453b58d3e2b661011513d