PP-2117: Rename KelvinService to ResultSinkService.
Summary:
Depends on D5927. This protobuf api breaking change should also still be ok because these services are only used internally in Vizier and we don't yet have rolling upgrades.
KelvinService isn't accurate anymore, it is more accurate to call it ResultSinkService because it is the api for any service that implements the apis necessary to receive a result chunk (aka row batch or exec stats). This group will include both Kelvin and Query broker in the near term.

Test Plan: existing

Reviewers: zasgar, michelle, philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5928

GitOrigin-RevId: dcfa43436f928daea8c9e0620a9ec9c0f056e3d5