Final Compiler and Planner fixes for partial aggregation
Summary:
proto serializer for blocking agg relies on relations being right, that wasn't the case here. Hack that fixese it for now but need clarification on usage before providing something better.

Also fixed the way we load aggregate expressions so that we load the functions correctly instead of trying to evaluate the children.

Test Plan: passes previously broken test and also fails in execution rather than in compilation for partial aggregates (which is expected, we want to update the aggregation operator).

Reviewers: zasgar, #engineering, nserrino

Reviewed By: #engineering, nserrino

Subscribers: nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D5921

GitOrigin-RevId: 1555dc1375116203070ddc4d23201741b9d1006f