Autocomplete: Add suggestions to empty tabstop
Summary:
when the user selects an entity, we automatically create a new, empty tabstop that we move their cursor to.
however, this tabstop is not populated with any suggestions, so the interaction seems a bit weird until the user starts typing again.

updated the backend so that if we do create new, empty tabstop, we also make a call to elastic to get a set of suggestions.

Test Plan: unit test, ran in cloud dev and checked that it did the right thing

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5941

GitOrigin-RevId: 5233d830c8a98362e3ef4da478220bd945a15a5b