Autocomplete: empty input should match all scripts
Summary:
until we have the script indexer up and ready, we've been matching scripts using the sahilm fuzzy search golang module.
this fuzzy search doesn't match any words when given an empty input string. in our case, we want it to match all scripts. these scripts will be filtered down further depending on the type of the input arguments already specified.
i updated the suggester code so that this is the case.
again, all of this script matching is just temporary until we have the script indexer... which probably won't be for a while

Test Plan: tried in cloud dev

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5945

GitOrigin-RevId: 7a7f8fecf4f5bbe0a8cb78ea8bd40c0da6cb2364