PP-2117: Implement TransferResultChunk logic on querybroker
Summary:
This is the endpoint that will be responsible for receiving result chunks from the agent.
It will subsume some of the logic in the ExecuteScript endpoint that is responsible for waiting for query results.

I moved mockContext to be a general testing utility (previously used only in one test).

Also, there was a name collision in the protobuf api, there were two objects being produced both called ExecutionAndTimingInfo (one with an underscore at the end).
Since this field isn't currently used anywhere, I renamed the message type to QueryExecutionAndTimingInfo to minimize confusion.
This is the lint-breaking protobuf API change in this diff.

Next diffs:
1. Implement query result forwarder
2. Hook query result forwarder into ExecuteScript
3. Change compiler to produce plans whose final sinks are GRPC sinks that write to the qb, rather than memory sinks.
4. Kelvin should not use ReceiveAgentQueryResult endpoint when all sinks are GRPC sinks
5. Deprecate ReceiveAgentQueryResult and stale code in ExecuteScript

Test Plan: added unit tests

Reviewers: philkuz, zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5944

GitOrigin-RevId: 2cbb7aaa5961439b18eab7bd31cf601c07d84398