PP-2117: Update handling of result schema for TransferResultChunk streamed results.
Summary:
Made some tweaks to the way that the output schema will be gotten from the query.
TransferResultChunkRequest will no longer send the output schema (this field was never hooked up to anything).
Instead, the GRPC Sinks that are not internal sinks will contain the full information about their output relation,
and those will be walked and collected (alongside memory sinks) and passed to the client using the existing code paths.
Needed to write a utility to convert the plan map to the output schema, so added a unit test for that as well.
A lot of this logic will be removed once ReceiveAgentResult is deprecated, but OutputSchemaFromPlan should continue to
be necessary and stick around.

Test Plan: unit test

Reviewers: zasgar, philkuz, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5960

GitOrigin-RevId: bee6d44fbe69e6788dec811d21c0b9f18da5c919