Cleanup vz updater interface in vzmgr
Summary:
we are planning to add an auto-updater in vzmgr.
before we can do that, there needs to be some cleanup done. i pulled out the logic for updating viziers into a new struct, which will be responsible for doing the auto-update in the future.

Test Plan: deployed to staging and deployed a new cluster + updated, to make sure it still works.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5964

GitOrigin-RevId: 921ec19b1043820bfb0759749c9d2536a362dec7