PP-2117: Result sinks should be GPRC sinks, not memory sinks.
Summary:
Depends on D6051. Depends on D6052.

This is the big change that switches over the compiler to generate GRPC sinks (which stream to the query broker) rather than memory sinks. It's a big change but most of it is test code (though I tried to separate out code into separate diffs when possible along the way).

Next up will be to clean up in Carnot the old handlers for the memory sink results being sent to the query broker. Then, the query broker itself needs to be refactored, since it buffers up results until EOS is received and then streams them back to the client.

Test Plan: edited tests

Reviewers: philkuz, jamesbartlett, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6053

GitOrigin-RevId: 78a51505103ad3038807a3c791dd0539d4a46bc2