Update job should apply only updated clusterrole first
Summary:
The initial time the update job applies the bootstrap YAML is so that the clusterroles get updated before running the following steps which may depend on the new clusterrole.
however, if something in the bootstrap YAML requires the updated clusterrole (such as podsecuritypolicy), depending on the ordering of how the YAML is applied, things may fail.
updated the apply YAML code so that we can apply only certain resourceTypes within that YAML. this will allow us to only apply the new clusterrole initially.
the rest of the bootstrap YAML will be applied at a later point, except now with the guarantee that the clusterrole is updated.

Test Plan: updated a vizier

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5970

GitOrigin-RevId: 3b3e7a9a15d045ba3b52db8b7798cf4a139c55b4