Fix bug where indexer was missing some resource versions
Summary:
noticed a bug that we weren't indexing all resources as expected.
upon closer inspection, it was a major bug with comparing resource versions.

when doing a range request on etcd, it gets the keys lexicographically.
if these were the keys on etcd: /test/1 /test/2 /test/20 /test/3
And we wanted to get the range: /test/1 to /test/20
it would just get /test/1, /test/2, but not /test/3 like we would want.

we need to pad the keys so that we actually get the correct numerical range.

once that was fixed, a similar bug surfaced in the metadata_reader for comparing resource versions.
since they are strings, it was also doing a lexicographic comparison. I opted to do a similar padding, then comparison method.

An alternative is to convert the resource versions to numbers and compare from there. but since resource version is a string in k8s, I worry about environments in the future where the resource version is not necessarily a number.

Test Plan: unit tests, ran cloud dev and checked to make sure all resource versions were actually sent.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5976

GitOrigin-RevId: 4335414f5afa981da276b78b745279ce16e4cb7f