Fix updates for rcs
Summary:
we recently changed our image path to be public for release viziers, and private for rc viziers.
that broke the updater job, because cloudconnector had the private path hardcoded. i made another change to use the path from the current cloudconnector image, but that was incorrect because it doesn't allow you to update from an rc -> public release and vice versa.
that was mostly because i didnt want to hardcode the image paths, but it looks like we'll need to. we should decide what the image path is based on whether the image tag has a "-" in it, which indicates that it is an rc

Test Plan: created an rc, tried to update it to another rc and tried to update it to an official release

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5980

GitOrigin-RevId: d80edcf2cb7403da541bb02cb0e9abe6961c245f