PP-2117: Add local GRPC result sink server to Carnot test utilities to receive streamed query results during end-to-end testing and local execution.
Summary: After PP-2117, all queries will be streaming and terminate in a GRPCSink. For local Carnot execution and end-to-end tests like benchmarks, we need some way of collecting those results at the end of the query, because they won't be placed in a memory sink anymore. This diff adds a GRPC server to receive those results and read them out later if needed. It integrates this Server into the blocking agg benchmark as an example (will integrate it into the rest afterwards).

Test Plan: n/a

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5983

GitOrigin-RevId: 2a0d1ef6d93084b8c65479ef7744b79ea6e4ee5e