C++ Argument Tracing
Summary:
This diff adds the C++ argument tracing feature.

It handles some of the complexities of arguments which get passed through registers vs through the stack. Some future work is required to complete the set of rules, but it handles moderate cases right now.

Test Plan: A moderately complex case added, and it traces the function properly.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5955

GitOrigin-RevId: 7b3d3ffe5e78641a00ec6bcca09cc92f275077c7