PP-2117: Query broker should pass its address to the compiler via the LogicalPlannerState in order to receive GRPC Sink results
Summary: GRPCSinks will be used to stream query results to the query broker after PP-2117. They need to know the address to send those results to. Adding to LogicalPlannerState a ResultAddress field that is passed by the query broker to the Plan() method. This will be used by the compiler to set the final address on those result GRPC Sinks.

Test Plan: updated tests

Reviewers: zasgar, philkuz, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5974

GitOrigin-RevId: bc9854165ec7037f6b644094a76cfa9a19be576f