Fix messaging when performing CLI auth for unregistered account
Summary:
Omid ran into this issue the other day when trying out pixie cloud dev. He hadn't realized his account wasn't yet registered with Pixie Cloud dev, and was confused when trying to authenticate with the CLI because it didn't give more info about why the token failed.
rather than exposing the generic "token cannot be obtained" error, I use the error included in the response body. this error is fine to expose, since it is the same error that we show when a user fails to login through the regular UI path.
this message is passed to the cli-auth-complete page, which used to always say that authentication completed successfully.
this message is now also displayed for manual mode.

Test Plan:
tried it with pixie cloud dev.
screenshots:
{F92105}

{F92106}

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5996

GitOrigin-RevId: 5e0acfad35a1888a7a4e6f03481b6f8fa6e3b076