PP-2117: Propagate result address for query results to the compiler state
Summary: Depends on D5974. The GRPCSinks produced by the compiler that are the final leaf nodes for a query need to know where to send their results to. This ddiff adds that field to the CompilerState so that it can be accessible by the compiler when producing those sinks.

Test Plan: existing

Reviewers: zasgar, philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D5979

GitOrigin-RevId: a0d2ab96b901aceddc6fff3c524b74e756e01df4