PP-2117: Add logic to Carnot to send exec stats over new streaming API
Summary:
In the old API, the exec stats are sent as part of the final batch result, and collected via the Done() API that will soon be deprecated.
In the new API, both row batches and exec stats alike will be sent via the same API. This diff adds in code to support that case, although it won't
currently be triggered until we move the result sinks from memory sources to grpc sinks in the compiler.

This change should also fix the error where the output rows / bytes are always 0 -- since we forgot the step where we add it across all of the agents we see.

Test Plan: unit

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6009

GitOrigin-RevId: d109b5f83ad9b40d7fc68294c1d59cd11df2bb08