Prevent vz auto-updater from getting stuck
Summary:
this code was refactored from another place which did have a ctx with a timeout. however, that is no longer the case and its possible for this for loop to wait indefinitely if either of the messages get lost in NATS somewhere.
if this happens, we should return after a certain amount of time and just retry again later.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6020

GitOrigin-RevId: f3eb74a8d7f0357ca1f1d9067f5011c485a2d236