PP-2117: Unify AddServiceTokenToClientToken logic and use it in GRPCSink to set bearer token when sending results externally.
Summary: We have logic to create and add service tokens in a few places in our CPP code already (when they send GRPC results to the query broker or when they make certain requests to the metadata service). We had it in 2 places, and now the GRPCSink class needs to use that same logic as well. As a result, in this change, I put that code all in one place (manager.cc) and then that function is pushed down through the contexts and states to anyone that needs it. This approach was chosen to ensure that auth logic highly specific to our codebase was not going to be linked directly into the carnot library.

Test Plan: added / edited existing, will run UDTFs on skaffold before landing

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6021

GitOrigin-RevId: 87563caa5fbae63550a21d335ad80dd68f060a42