Make autocomplete args optional
Summary:
before we always required them to be filled out. however, we don't have a notion of which args are/aren't required at the moment without actual script indexing.
instead we should make the args optional. if the arg is empty, we should just consider the command valid and executable

Test Plan: ran in plc-dev

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6023

GitOrigin-RevId: db6634d8f69b874fff0d0e190929dbf5912e6bc2