[PP-2138] CPP Docstring extractor
Summary: The extractor grabs all the docs from the QLObjecdts and UDFs. The UDF part is still in progres (cc @zasgar) but the qlobject pipeline is running. This just returns the docstrings. They need to be parsed into a structured format that pixie-docs rendersd.

Test Plan: Wrote tests and used in coordination with parser, which will come in follow up diff

Reviewers: zasgar, nserrino, jamesbartlett, #engineering

Reviewed By: nserrino, #engineering

Subscribers: zasgar

JIRA Issues: PP-2138

Differential Revision: https://phab.corp.pixielabs.ai/D6003

GitOrigin-RevId: ff48a7e5a811c4495917c1f57938f67eae5aa81e