DynamicTracing: StructBlob for tracing entire types
Summary:
Before we can trace interfaces, we need a way of tracing struct types. This diff is the first part in that direction.

Part 2 will focus on decoding the struct blob into JSON.

Test Plan: A placeholder end-to-end test was added for tracing structs. Confirms that the probe can deploy and the length of the struct is recorded. Once part 2 is added, the test will be beefed up.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6002

GitOrigin-RevId: 70b1a6e4e24ae9ebb129e08067a3274d25bf363f