PP-2130 Fix autocomplete slow backspace
Summary:
We were not handling backspaces the same way we were handling insertions. this is because we were still relying on  making an API call per backspace.
to avoid this, we need to correctly update our tabStop structure according to the deletion. this way, we don't have to rely on parsing the tabstop structure that we get from the backend.

Test Plan: unit test, ran webpack and tried backspaces in different cases

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6029

GitOrigin-RevId: 26fd6acbf5c634378359cd4c113dc9f241e0db0c