PP-2117: Make SSL target name for query results configurable.
Summary: Currently in our Carnot instances, we hardcode an ssl target override to "kelvin.pl.svc" when sending query results from pems to kelvins. With the addition of query results being sent from kelvins to the query broker, that ssl target override will now either be kelvin.pl.svc or vizier-query-broker.pl.svc. To support this variance and also cases where it may be different in the future (or different between customers), I am adding it to the logical distributed state, where it will be put in the GRPCSink plan output as a "connection option" for which ssl target name to use. This is diff 1 of either 2 or 3 of these changes, starting at the go/query broker level. next will be plan, and then finally exec nodes to use the configurable ssl target name.

Test Plan: edited existing

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6032

GitOrigin-RevId: 61047bf3b12b5ae46321930cdbac236e462903eb