PP-2117 Make SSL target name for query results configurable, part 2.
Summary: Depends on D6032. The compiler/planner side of the changes to make SSL target name for query results configurable, so that we can support vizier-query-broker.pl.svc and kelvin.pl.svc, as well as future targets that may differ in the future (rather than hardcoded in the manager).

Test Plan: edited/existing

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6033

GitOrigin-RevId: 8e747e5f09c192410366ab6fabcd0d47d9705e1a