Add client side streaming test for grpc server test
Summary:
I was trying to debug a grpc authentication issue, and added this unit test to make sure that the problem wasn't in this component.
It turns out the problem must be somewhere else, but we might as well check this in if we have it.

Test Plan: the test

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6048

GitOrigin-RevId: e4da7ab27b0f28f9b4e14bf1aa6f94f046d2eb84