PP-2117: some planner and engine preparations for turning memory sinks into grpc sinks.
Summary: The diff that switches memory sinks over to grpc sinks in the planner is coming soon and it is fairly large. This diff is all of the odds and ends in that change that can be separated out for that diff, for diff size reduction purposes. These changes all enable either the switch of GRPC sinks instead of memory sinks as the output sinks in query plans, or enhance testing for those changes.

Test Plan: existing

Reviewers: michelle, philkuz, zasgar, #engineering, jamesbartlett

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6052

GitOrigin-RevId: f49ac7210ab8ea8830cf579fc1ba428828f63939