Improve autocomplete namespace matching
Summary:
before we were only matching namespaces if the user explicitly typed "<namespace>/" in the input. it would also only match the namespace text exactly.
instead, it should be a fuzzy-match like the name, and we should try to match with namespaces even if the user hasn't explicitly typed the "/"

Test Plan: ran plc-dev

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6054

GitOrigin-RevId: 7f1ac8a3e7080d818af3e9fe7549af08dbdb7a1a