Show org-specific scripts in autocomplete
Summary:
the current autocomplete doesnt show org-specific scripts. this is mostly because the script-autocompleting logic is temporary until we have an actual script indexer. for now, we are just using the bundle reader. a script in the bundle reader has an orgName for each script. however, we only can get the orgID from the jwt when autocompleting. one option is to add the orgName to the jwt, but that is redundant information.
another option is to make a request for the orgName, given the orgID, for every autocomplete call. but we don't want to slow down autocomplete with redundant requests.
instead, what I decided is to make a map of orgID -> orgName upon initialization of the API service. the suggester can then use this mapping to determine whether to show a script, given the orgID.
again, this is all temporary until we have script indexing.

Test Plan: ran plc-dev, checked that i could see pixielabs.org specific scripts

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6057

GitOrigin-RevId: 83c6385cec8a0c7d138cae7b9e8e11bfb0e746cd