PP-2117: Introduce new QueryResultForwarder to support streaming Kelvin responses back from QueryBroker to client.
Summary:
This is the new QueryResultForwarder, which in the next diff will take over a lot of the work that the QueryExecutor used to do.
The QueryResultForwarder is built to handle streaming responses from the agent and forward those responses (transformed to the public proto)
to the client stream. Next diff will change the query broker server to integrate the new QueryResultForwarder.

Test Plan: added various tests, including cancellation, to check for hangs

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2117

Differential Revision: https://phab.corp.pixielabs.ai/D6066

GitOrigin-RevId: ec1b6a82b63e713c45f8eb5fe49138c16c19482a