Update UI to fetch both opensource + px scripts
Summary:
made two copies of the bundle.json as named them as bundle-px.json and bundle-os.json for testing.
this just updates the UI so it fetches both px and open-source bundles in a Promise.all.

The pxl dev server should operate as normal, unless we also want the pxl dev server to host opensource scripts as well. that would just be a small modification to the current code.

Test Plan: ran webpack

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6085

GitOrigin-RevId: 5b3fb7f8034d5a97497d799084f42631c0f333bc