Handle timeouts better in update job
Summary:
we saw a case in customer where deleting objects took longer than the 5 min timeout, which caused the updater job to just exit.
instead, if theres a timeout error, we should try to proceed as normal. it should be OK to apply a YAML when a previous one is still deleting (unless deleting a namespace, which we aren't in this case).
also made the timeout a little longer, which should help. however, k8s pods can be known to take hours to terminate sometimes.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6099

GitOrigin-RevId: 20bb25ef8c583bb1b3c0a9193cb23f62317e0853