Fix issue with esutils/Index where non-updates to settings would cause failures.
Summary: Some elastic settings can't be changed on an open index, so if there is an attempt to change them this will cause a failure. But before this change, even if the setting didn't change the update was still being called, leading to the same failure.

Test Plan: Added a test for this failure case, test passes.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6108

GitOrigin-RevId: d5ee23c7a226dd40ca4e99026f0d08157a284660