PP-2115: Add a heartbeat query result message to carnotpb to send to keep streaming queries alive.
Summary:
added handling in the query broker for these heartbeats, and add some unit tests that were missing.
this will help for streaming queries with sparse data where data isn't being sent over that often but we don't want to timeout.
we can use the heartbeats to send over the execution stats for the query up to that point as well, so that streaming queries
can still show something for the execution stats.

Test Plan: added tests

Reviewers: michelle, zasgar, #engineering, philkuz

Reviewed By: #engineering, philkuz

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6113

GitOrigin-RevId: d4df2b1e4da244425b98ccd30cc16a620908ae5d