Periodically log etcd status
Summary:
currently we have no insight into the etcd running on our customer's clusters.
luckily, with the client we can hit some of the etcd endpoints to see the amount of space we're currently using.
we're constantly running into space issues on customer, so this should give us a sense of how often we may need to defrag or if a defrag won't be enough.

Test Plan: ran in skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6115

GitOrigin-RevId: cd3756359cf209ccfd3d11df28726e6959fd98e5