PP-2178 Fix metadata agent data inconsistency
Summary:
We ran into an error in the querybroker where the agent state was unable to update because of this error occurring every 5s:
```Received error running agent tracker loop. Retrying in 5 seconds. Could not update agent table metadata of unknown agent 7309788f-79d8-4f80-ad7d-b6c3e13b47aa```

This means that the metadata sent the qb schema/data for an agent which has been deleted.
Taking a look at the code, it is definitely possible to update schema/data for an agent which no longer exists.
since we process agent updates in a queue, its possible for us to put an update on the queue and not process it before the agent has already been deleted. this is especially possible in larger clusters where there may be many updates in the queue.
before updating the datastore with the new schema/data, we should check whether the agent actually exists.

Test Plan: unit test

Reviewers: nserrino, #engineering, philkuz

Reviewed By: #engineering, philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D6116

GitOrigin-RevId: 37b1eab1cd03cdf796d701e35d951fc8f5e7e582