Fix bug with limit producing empty row batches.
Summary:
A while back, I was trying to fix a bug with Limit nodes when there were 2 separate graphs one limit would prevent the other graph from running. In fixing that bug, I introduced a new bug with limits that have multiple sources (i.e. a limit after a union), where the limit would output empty row batches for each of the other sources if one of the sources was enough to reach the limit. (I discovered this b/c these empty row batches each of eos set, causing an agg after the limit to output 1 row for each row batch)

This reverts my old fix and adds a new fix that works in both cases.

Test Plan: I added tests for both cases, as expected the second case's test fails on master, and both pass with this diff.

Reviewers: #engineering, philkuz

Reviewed By: #engineering, philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D6118

GitOrigin-RevId: b20f695ef286c14880b765e791bb1bc27ea62ff4