Fix updater timeout issue
Summary:
turns out the previous fix didn't work, because the error wasn't actually the type I expected it to be.
turns out theres no specific error type that we can check for this timeout error, so we have to check it manually.

Test Plan: updated the job timeout to be 2s instead of 10 mins. confirm that my cluster updated properly anyway

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6112

GitOrigin-RevId: 5eaebfd5049f1ff5ab9e4badce9ea8d19ff7c406