PP-2175 Fix bug for registering viziers with the same name
Summary:
when registering a vizier with a name that we already have in the database, we do a loop 10 times to try registering with the name <name_%d>, where d is the number of times the loop has run.
this works when there are only 10 clusters with the same name, but once the 11th one tries to register, we error. obviously this isnt scalable. instead, we can try to get a random number from a large pool of large numbers.
this is happening especially in the case where everyone's cluster is named "minikube"

Test Plan: unit test

Reviewers: zasgar, #engineering, oazizi

Reviewed By: #engineering, oazizi

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D6107

GitOrigin-RevId: 19cab69e994285f865f7e1f11c2849552140073b