Update job: Keep cloud-connector alive
Summary:
We want to keep the cloud connector alive as much as possible, so that we always have a way to recover without requiring a fresh reinstall from our customers.
The current update flow is:
- delete all vizier resources (this includes cloudconn + its deps, minus etcd/nats)
- launch new versions of all vizier resources
In between the time of deletion and launching, the cloud connector deployment is completely gone from the namespace. if something hapepns in that time, it is possible that we may never get the cloud connector back.

Instead, the flow is updated to be more like this:
- delete all vizier resources (minus cloudconn + its deps)
- launch new versions of all vizier resources. for cloudconn + its deps, this is an update.
Now, there is no period of time where there is no cloud connector deployment.

We need to additionally bounce the cloudconnector pod to handle a case where we try to update to the same version we're already on. since this is an update now, the cloudconnector pod just keeps running. however, we expect the cloudconnector to clean up the update job upon startup.

Test Plan:
tried 4 cases:
- update 0.4.5 to RC with new update changes
- deploy RC with update changes
- update RC to same RC version
- update RC to newer RC version

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6129

GitOrigin-RevId: 365e9e74707e5d89a7f9c8794f8830635cf524fb